Add unit tests
I have added unit tests for both the Dateful class and the factories that
produce instances of it. These tests are written with the testing
framework AVA, and cover all the public methods.

I have also shuffled the project structure around to hold the tests, and
rewritten parts of both original files in response to issues the tests
uncovered.