Setup build pipeline
With this commit the build pipeline is complete. Webpack has been included
and configured, and the awesome-typescript-loader for it is set up. The
command `yarn run bundle` will now create a dist/ folder containing a
package.json, an index.js bundle file (with sourcemap) and a dateful.d.ts
definition file. We're now ready to publish an initial version to npm!