Add method cancelTimezoneOffset
This new method cancels out the effect of the local timezone by
subtracting the offset from the time. This is important in certain
use-cases, such as when a UI datepicker is used in the browser to pick
only the date part (setting the time part automatically to midnight in the
local timezone), but due to the timezone offset the date sent to the
server is actually the previous day (at, say 11pm).