Move timezone offset fix into constructor
I have moved the triggering of the timezone offset fix into the
constructor instead of allowing it to be in its own method. This ensures
that the fix is applied the moment the Dateful is created, before timeless
can trigger and wipe out the time component.

The constructor now has a new optional parameter, fixTimezoneOffset, in
order to trigger this behaviour.